+++ release-notes.md
@@ -1,6 +1,9 @@
 Pending changes
 ===============
 
+Changes:
+- Added "export to NIfTI" button for 3D series and instances.
+
 Fixes:
 - Fix [missing fields when browsing remote DICOMweb servers at series level](https://discourse.orthanc-server.org/t/dicom-web-client-not-requesting-all-tags-it-intends-to-display/6348)
 - Fix #95 [DICOMweb: Instance retrieve sends wrong WADO-RS request](https://github.com/orthanc-server/orthanc-explorer-2/issues/95)